// apps/git-client/src/generated/SubmoduleDiff.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SubmoduleDiff = { path: string, beforeOid: string | null, afterOid: string | null, beforeSubject: string | null, afterSubject: string | null, ahead: number | null, behind: number | null, };